add data Fields and other modifications

+++ Models/Employee.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using EmployeeDemo.Extensions;
 using System.ComponentModel.DataAnnotations;
 using System.Collections.Generic;
 
@@ -10,7 +9,7 @@ namespace EmployeeDemo.Models
         [Key]
         public int Id { get; set; }
         public string FirstName { get; set; }
-        public int Phone { get; set; }
+        public long Phone { get; set; }
         public DateTime DateOfBirth { get; set; }
         public int Age { get; set; }
         public string Department { get; set; }
